WIP: move check_file to common module
