+++ aodntools/timeseries_products/aggregated_timeseries.py
@@ -13,7 +13,7 @@ from netCDF4 import Dataset, num2date, stringtochar
 from pkg_resources import resource_filename
 
 from aodntools import __version__
-from aodntools.timeseries_products.common import NoInputFilesError
+from aodntools.timeseries_products.common import NoInputFilesError, check_file
 
 TEMPLATE_JSON = resource_filename(__name__, 'aggregated_timeseries_template.json')
 
@@ -32,71 +32,6 @@ def sort_files(files_to_agg, input_dir=''):
             time_start.append(np.datetime64(ds.time_deployment_start.rstrip('Z')))
     tuples = sorted(zip(time_start, files_to_agg))
     return [t[1] for t in tuples]
-
-
-def check_file(nc, site_code, VoI):
-    """
-    Return list of errors found in the file.
-    Checks applied:
-    * Correct site_code
-    * Variables of interest is present
-    * Variables TIME, LATITUDE and LONGITUDE are present
-    * NOMINAL_DEPTH is present as variable or attribute
-    * file_version is FV01
-    * if LATITUDE or LONIGUTDE are dimension, they have length 1
-    * Global attributes time_deployment_start and time_deployment_end exist
-
-    :param nc: xarray dataset
-    :param site_code: code of the mooring site
-    :param VoI: string. Variable of Interest
-    :return: dictionary with the file name and list of failed tests
-    """
-
-    attributes = list(nc.attrs)
-    variables = list(nc.variables)
-    allowed_dimensions = ['TIME', 'LATITUDE', 'LONGITUDE']
-    error_list = []
-
-    if site_code != nc.site_code:
-        error_list.append('Wrong site_code: ' + site_code)
-
-    nc_file_version = nc.file_version
-    if 'Level 1' not in nc_file_version:
-        error_list.append('Wrong file version: ' + nc_file_version)
-
-    if 'TIME' not in variables:
-        error_list.append('TIME variable missing')
-
-    if 'LATITUDE' not in variables:
-        error_list.append('LATITUDE variable missing')
-
-    if 'LONGITUDE' not in variables:
-        error_list.append('LONGITUDE variable missing')
-
-    if VoI not in variables:
-        error_list.append(VoI + ' not in file')
-    else:
-        VoIdimensions = list(nc[VoI].dims)
-        if 'TIME' not in VoIdimensions:
-            error_list.append('TIME is not a dimension')
-        if 'LATITUDE' in VoIdimensions and len(nc.LATITUDE) > 1:
-            error_list.append('more than one LATITUDE')
-        if 'LONGITUDE' in VoIdimensions and len(nc.LONGITUDE) > 1:
-            error_list.append('more than one LONGITUDE')
-        for d in range(len(VoIdimensions)):
-            if VoIdimensions[d] not in allowed_dimensions:
-                error_list.append('not allowed dimensions: ' + VoIdimensions[d])
-                break
-
-    if 'NOMINAL_DEPTH' not in variables and 'instrument_nominal_depth' not in attributes:
-        error_list.append('no NOMINAL_DEPTH')
-
-    if 'time_deployment_start' not in attributes:
-        error_list.append('no time_deployment_start attribute')
-    if 'time_deployment_end' not in attributes:
-        error_list.append('no time_deployment_end attribute')
-
-    return error_list
 
 
 def get_variable_values(nc, variable):
